Fix minor errors in client, server, and queue and rename egyptainratscrew
Contributes to #16, #19
