+++ CardGames/games/egyptianratscrew.py
